cleanup graph-model.ts, move abbreviateLargeNumber to units.ts
